update: change userRepository PK to Long

+++ src/main/java/com/ceos18/springboot/user/repository/UserRepository.java
@@ -1,13 +1,14 @@
 package com.ceos18.springboot.user.repository;
 
 import com.ceos18.springboot.user.entity.User;
+import org.apache.catalina.mapper.Mapper;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByNickname(String nickname);
     Optional<User> findByEmail(String email);
